Fixed a few more logic flow holes.
► Fixed logic error when the script is first initialized, it will prompt the
user because "Current Commit Notes.txt" is unchaged/empty. Now it will check if
any of the 2 prerequisite files are present for it to operate before proceeding
with the rest of the script.
► Fixed "print_last_logs" method where -C is pointed at a specific directory
and not the current project directory.
► Fixed "check_git_status" method where it can only detect a clean working
tree but not a repository with no commits yet.
► Added option to either exit or change current project directory when script
detects the current repository's working tree is clean instead of exiting
immediately.